fix just percent test. valid now

+++ general/includes/size_parser.h
@@ -5,6 +5,6 @@
 #ifndef PRINTF_SIZE_PARSER_H
 #define PRINTF_SIZE_PARSER_H
 
-ssize_t 	parse_size(const char *format, t_spec *spec);
+int parse_size(const char *format, t_spec *spec);
 
 #endif //PRINTF_SIZE_PARSER_H
